different currencies accepted
